move apeals objects and add appeaks manager
