+++ Freelance_WbSellerFeedbackPareser/WbAppealsManager/WbAppealReader.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json.Linq;
 
-namespace Freelance_WbSellerFeedbackPareser
+namespace Freelance_WbSellerFeedbackPareser.WbAppealsManager
 {
     internal class WbAppealReader
     {
