+++ Freelance_WbSellerFeedbackPareser/WbAppealsManager/WbAppealsExcelCreator.cs
@@ -1,6 +1,6 @@
 ﻿using ClosedXML.Excel;
 
-namespace Freelance_WbSellerFeedbackPareser
+namespace Freelance_WbSellerFeedbackPareser.WbAppealsManager
 {
     internal class WbAppealsExcelCreator
     {
